requirements.txt
# Core dependencies
PySide6>=6.6.0
opencv-python-headless>=4.8.0
mediapipe>=0.10.8
numpy>=1.24.0
face_recognition>=1.3.0  # Biometric face verification
dlib>=19.24.0  # Required by face_recognition

# Windows API (kiosk mode)
pywin32>=306; sys_platform == 'win32'

# Networking
httpx>=0.25.0
python-dotenv>=1.0.0

# Audio processing
PyAudio>=0.2.14
webrtcvad>=2.0.10

# Cryptography
cryptography>=41.0.0

# Local database
# SQLite is built-in

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Packaging
pyinstaller>=6.0.0

# Utilities
pillow>=10.0.0
python-dateutil>=2.8.2
